% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{count_rc}
\alias{count_rc}
\title{Calculate and plot RC}
\usage{
count_rc(dt_file, metadata_file, output_dir = NULL)
}
\arguments{
\item{dt_file}{Data table file}

\item{metadata_file}{Metadata table file}

\item{output_dir}{Output dir}
}
\value{
Numeric vector
}
\description{
Calculate correlation to reference datasets and plot related scatter plot
}
\examples{
count_rc(dt = sample_data, metadata = sample_metadata)

}
