% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{count_recall}
\alias{count_recall}
\title{count_recall}
\usage{
count_recall(dt_file, metadata_file)
}
\arguments{
\item{dt_file}{Data table file}

\item{metadata_file}{Metadata table file}
}
\value{
Numeric vector
}
\description{
Calculate Recall of DAMs
}
