% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{count_snr}
\alias{count_snr}
\title{Calculate SNR and plot PCA}
\usage{
count_snr(dt_file, metadata_file, output_dir = NULL)
}
\arguments{
\item{dt_file}{Data table file}

\item{metadata_file}{Metadata table file}

\item{output_dir}{Output dir}
}
\value{
Numeric vector
}
\description{
Plot PCA and calculate SNR based on the first two principal componants of PCA.
}
