% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountRecall.R
\name{CountRecall}
\alias{CountRecall}
\title{CountRecall}
\usage{
CountRecall(
  dt.path = NULL,
  metadata.path = NULL,
  output.path = NULL,
  dt = NULL,
  metadata = NULL
)
}
\arguments{
\item{dt}{Data table}

\item{metadata}{Data table}
}
\value{
Numeric vector
}
\description{
Calculate Recall of DAMs
}
\examples{
CountRecall(dt=sample_data,metadata=sample_metadata)

}
