% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSE.R
\name{CountRMSE}
\alias{CountRMSE}
\title{Calculate and RMSE}
\usage{
CountRMSE(
  dt.path = NULL,
  metadata.path = NULL,
  output.path = NULL,
  dt = NULL,
  metadata = NULL
)
}
\arguments{
\item{dt}{Data table}

\item{metadata}{Data table}
}
\value{
Numeric vector
}
\description{
Calculate RMSE of ratios of spike-in pairs
}
\examples{
CountRMSE(dt=sample_data,metadata=sample_metadata)

}
