% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{detect_gaps}
\alias{detect_gaps}
\title{detect_gaps}
\usage{
detect_gaps(df, expected_interval = 30, date_field = "DATECT")
}
\arguments{
\item{df}{A data frame of met data}

\item{expected_interval}{Expected time interval of series, Default: '30 min'}

\item{date_field}{Column name for POSIX date/time variable in df, Default: 'DATECT'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Detects any gaps in a data frame representing a time series,
ie. find any intervals greater than the expected interval.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 gaps <- detect_gaps(l_logr$df)
 }
}
}
