% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_for_ceda.R
\name{format_for_ceda}
\alias{format_for_ceda}
\title{Format the data for submission to CEDA}
\usage{
format_for_ceda(
  l_lev,
  v_names = c("TA_4_1_1", "RH_4_1_1", "PA_4_1_1", "RG_4_1_0", "PPFD_IN_4_1_1",
    "RN_5_1_1", "P_12_1_1", "TS", "G", "SWC", "WS_6_1_1", "WD_6_1_1", "D_SNOW", "WTD")
)
}
\arguments{
\item{l_lev}{A named list with two components \code{df} and \code{df_qc},
which are the data frames of the observations and the associated quality
control codes.}

\item{v_names}{A character vector of column names in the data frame \code{df}
that should be included in the new data frame.}
}
\value{
A new data frame with the selected columns from \code{df} and
\code{df_qc}.
}
\description{
Format the data for submission to CEDA
}
