% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{generate_met_report}
\alias{generate_met_report}
\title{Generate Quartet Metabolomics report}
\usage{
generate_met_report(
  qc_result,
  report_template,
  report_dir = NULL,
  report_name = NULL
)
}
\arguments{
\item{qc_result}{list}

\item{report_template}{character}

\item{report_dir}{character}

\item{report_name}{character}
}
\description{
Use calculated Met result to generate report
}
\examples{
sample_data <- system.file("extdata", "sample_data.csv", package = "metqc")
sample_metadata <- system.file("extdata", "sample_metadata.csv", package = "metqc")

# Get performance report for metabolomics data
met_result <- get_performance(dt_file = sample_data, metadata_file = sample_metadata)

# 指定包内文档的路径
report_template <- system.file("extdata", "quartet_template.docx", package = "metqc")

# 运行函数
generate_met_report(qc_result = met_result, report_template = report_template)

}
