% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_report.R
\name{GenerateReport}
\alias{GenerateReport}
\title{Generate Quartet Metabolomics report}
\usage{
GenerateReport(Met_result = NULL, doc_file_path = NULL, output_path = NULL)
}
\arguments{
\item{Met_result}{list}

\item{output_path}{character}

\item{temp_doc_path}{character}
}
\value{
word file
}
\description{
Use calculated Met result to generate report
}
\examples{
# 加载示例 Met_result 对象
met_result_path <- system.file("extdata", "Met_result_example.RData", package = "MetQC")
load(met_result_path)

# 指定包内文档的路径
doc_file_path_example <- system.file("extdata", "Quartet_temp.docx", package = "MetQC")

# 假设输出路径是当前工作目录
output_path_example <- getwd()

# 运行函数
GenerateReport(Met_result=met_result, doc_file_path=doc_file_path_example, output_path=output_path_example)


}
