% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{get_performance}
\alias{get_performance}
\title{Get output table}
\usage{
get_performance(dt_file, metadata_file, output_dir = NULL)
}
\arguments{
\item{dt_file}{Data table file}

\item{metadata_file}{Metadata table file}

\item{output_dir}{Output dir}
}
\value{
Numeric vector
}
\description{
Get information of overall performance with historical data.
}
