% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistoricalData.R
\docType{data}
\name{HistoricalData}
\alias{HistoricalData}
\title{Historical QC data of metabaolomics}
\format{
A \code{data.table} with 4 columns.
}
\usage{
HistoricalData
}
\description{
A \code{data.table} file.
}
\keyword{datasets}
