% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute}
\alias{impute}
\title{impute}
\usage{
impute(
  y,
  l_met = l_met,
  method = "era5",
  qc_tokeep = 0,
  selection = TRUE,
  date_field = "DATECT",
  k = 40,
  fit = TRUE,
  x = NULL,
  df_era5 = NULL,
  lat = 55.792,
  lon = -3.243,
  plot_graph = TRUE
)
}
\arguments{
\item{y}{Response variable with missing values to be replaced
(variable name as a "quoted string")}

\item{l_met}{List of two data frames containing data and qc codes. Default: l_met}

\item{method}{Method to use for imputing missing values, Default: "era5"}

\item{qc_tokeep}{Which QC codes to leave unaltered when selecting values to impute, Default: 0 (raw data)}

\item{selection}{Denotes the points selected interactivelt by user in app Default: TRUE (= all)}

\item{date_field}{Name of the date field or variable in the data frame. Default: DATECT}

\item{k}{Number of knots to use when imputing using a GAM in the "time" method. Higher values give more flexibility = more wiggliness. Default: 40}

\item{fit}{Whether to fit a linear model or directly replace missing y with
x values when using either "regn" or "era5" methods, Default: TRUE}

\item{x}{The covariate with which to fit a linear model in the "regn" method. Default: NULL}

\item{df_era5}{The name of the data frame containing the corresponding ERA5 data. Default: df_era5}

\item{lat}{Latitude of the site for calculating day/night-time in "nightzero" method. Default: 55.792 (= Auchencorth)}

\item{lon}{Longitude of the site for calculating day/night-time in "nightzero" method. Default: -3.243 (= Auchencorth)}

\item{plot_graph}{Whether to produce a ggplot graphic - can be slow for large data sets. Default: TRUE}
}
\value{
List of two data frames containing data and qc codes with imputed values.
}
\description{
Impute missing values using various methods
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()) {
 #EXAMPLE1
l_met <- list(df = df, df_qc = df_qc)
l_met <- impute(y = "SW_IN", x = "PPFD_IN",  l_met)
 }
}
}
