% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_hmdb_id}
\alias{map_hmdb_id}
\title{Pre-process dataset}
\usage{
map_hmdb_id(dt_file)
}
\arguments{
\item{dt_file}{Data table file}
}
\value{
Numeric vector
}
\description{
Map the name of metabolites to the existing ID files.
}
