% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapIDs.R
\name{MapIDs}
\alias{MapIDs}
\title{Pre-process dataset}
\usage{
MapIDs(dt.path = NULL, dt = NULL)
}
\arguments{
\item{dt}{Data table}
}
\value{
Numeric vector
}
\description{
Map the name of metabolites to the existing ID files.
}
\examples{
MapIDs(dt=sample_data)

}
