% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metdb-package.R
\docType{package}
\name{metdb-package}
\alias{metdb}
\alias{metdb-package}
\title{metdb: Quality Control Of Field Site Meteorological Data}
\description{
Quality control of field site meteorological data.
}
\author{
\strong{Maintainer}: Thomas Zwagerman \email{thowag@ceh.ac.uk}

Authors:
\itemize{
  \item Pete Levy \email{plevy@ceh.ac.uk} (\href{https://orcid.org/0000-0002-8505-1901}{ORCID})
  \item David Leaver \email{leav@ceh.ac.uk}
}

}
\keyword{internal}
