% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetInfo.R
\docType{data}
\name{MetInfo}
\alias{MetInfo}
\title{Metabolite name mapping list.}
\format{
A \code{data.table} with 2 columns. 
\describe{
\item{HMDBID}{A column indicates the IDs of metabolites.}
\item{metabolites}{A column indicates the names of metabolites.}
}
}
\usage{
MetInfo
}
\description{
A \code{data.table} file with 2 columns.
}
\keyword{datasets}
