% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_colmap.R
\name{mod_colmap_server}
\alias{mod_colmap_server}
\title{Column Mapping Server}
\usage{
mod_colmap_server(id, uploaded, dictionary_list)
}
\arguments{
\item{id}{Module ID}

\item{uploaded}{Reactive containing a list with \verb{$data}}

\item{dictionary_list}{List of tibbles (preloaded dictionaries)}
}
\description{
Column Mapping Server
}
