% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload.R
\name{mod_upload_server}
\alias{mod_upload_server}
\title{Upload Module Server}
\usage{
mod_upload_server(id)
}
\arguments{
\item{id}{shiny id}
}
\value{
reactive list with $data (uploaded table)
}
\description{
Upload Module Server
}
