% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_heatmap_calendar}
\alias{plot_heatmap_calendar}
\title{Custom plotting function to construct a heatmap calendar}
\usage{
plot_heatmap_calendar(input_variable, df_qry)
}
\arguments{
\item{input_variable}{}

\item{df_qry}{}
}
\value{

}
\description{
Custom plotting function to construct a heatmap calendar
}
