% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample data for analysis}
\format{
A \code{data.table} with 14 columns. 
\describe{
\item{HMDBID}{A column indicates the ids of metabolites.}
}
}
\usage{
sample_data
}
\description{
A \code{data.table} file with columns named after different samples for testing.
}
\keyword{datasets}
