% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{generate_met_report}
\alias{generate_met_report}
\title{Generate Quartet Metabolomics report}
\usage{
generate_met_report(
  qc_result,
  report_template,
  report_dir = NULL,
  report_name = NULL
)
}
\arguments{
\item{qc_result}{list}

\item{report_template}{character}

\item{report_dir}{character}

\item{report_name}{character}
}
\description{
Use calculated Met result to generate report
}
\examples{
# 加载示例 qc_result 对象
qc_result_path <- system.file("extdata", "qc_result_example.RData", package = "metqc")
load(qc_result_path)

# 指定包内文档的路径
report_template <- system.file("extdata", "quartet_template.docx", package = "metqc")

# 运行函数
generate_met_report(qc_result = met_result, report_template = report_template)

}
