% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metqc-package.R
\docType{package}
\name{metqc-package}
\alias{metqc}
\alias{metqc-package}
\title{metqc: Quality Control Of Field Site Meteorological Data}
\description{
Quality control of field site meteorological data.
}
\author{
\strong{Maintainer}: Thomas Zwagerman \email{thowag@ceh.ac.uk}

Authors:
\itemize{
  \item Peter Levy \email{plevy@ceh.ac.uk} (\href{https://orcid.org/0000-0002-8505-1901}{ORCID})
  \item David Leaver \email{leav@ceh.ac.uk}
}

Other contributors:
\itemize{
  \item Juan Pablo Lobo Guerrero \email{jualob@ceh.ac.uk} [contributor]
  \item Karen Yeung \email{karung@ceh.ac.uk} [contributor]
}

}
\keyword{internal}
