% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_heatmap_calendar}
\alias{plot_heatmap_calendar}
\title{Custom plotting function to construct a heatmap calendar}
\usage{
plot_heatmap_calendar(input_variable, df_qry)
}
\arguments{
\item{input_variable}{might not need this anymore, given all variables need checking for the same date}

\item{df_qry}{the dataframe of the query, in order to extract the dates}
}
\value{

}
\description{
Custom plotting function to construct a heatmap calendar
}
