% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_heatmap_calendar}
\alias{plot_heatmap_calendar}
\title{plot_heatmap_calendar}
\usage{
plot_heatmap_calendar(df)
}
\arguments{
\item{df}{A data frame of met data}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Plots a calendar showing who has checked the data by date
}
\details{
Custom plotting function to construct a heatmap calendar

DETAILS
}
