% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotting_function}
\alias{plotting_function}
\title{Custom plotting function for each variable}
\usage{
plotting_function(input_variable)
}
\arguments{
\item{input_variable}{}
}
\value{

}
\description{
Custom plotting function for each variable
}
